Start services in forks and maintain their states
This commit does not care about dependencies yet, although it
acknowledges their existence in a slot.

All services are properly started in a fork(), while the parent does
what it can to maintain their state. A thread is used to run the
waitpid(), which will wait for the child to die.

Multithread safety will have to be handled. Whenever a change or a read
is happening to the service, a semaphore has to be held in order to
maintain the consistency of all the slots. This is not done yet.